Bluetooth: Fix GATT CEP not using 16 bits
Characteristic Extended Properties has 16 bits not 8 bits as value.


Former-commit-id: 66d6cf77cef243db395e8eb7c70f69755786a147

+++ include/bluetooth/gatt.h
@@ -48,7 +48,7 @@ DECL|flush|member|int (*flush)(const bt_addr_le_t *peer,
 DECL|handle|member|uint16_t handle;
 DECL|peer|member|bt_addr_le_t peer;
 DECL|perm|member|uint8_t perm;
-DECL|properties|member|uint8_t properties;
+DECL|properties|member|uint16_t properties;
 DECL|properties|member|uint8_t properties;
 DECL|read|member|int (*read)(const bt_addr_le_t *peer,
 DECL|start_handle|member|uint16_t start_handle;
